--- a/WARP.md
+++ b/WARP.md
@@ -1,0 +1,1 @@
+CLAUDE.md