--- a/python/src/agents/base_agent.py
+++ b/python/src/agents/base_agent.py
@@ -214,8 +214,8 @@ class BaseAgent(ABC, Generic[DepsT, OutputT]):
                 timeout=120.0,  # 2 minute timeout for agent operations
             )
             self.logger.info(f"Agent {self.name} completed successfully")
-            # PydanticAI returns a RunResult with data attribute
-            return result.data
+            # PydanticAI returns a RunResult with output attribute
+            return result.output
         except asyncio.TimeoutError:
             self.logger.error(f"Agent {self.name} timed out after 120 seconds")
             raise Exception(f"Agent {self.name} operation timed out - taking too long to respond")